git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@3615 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/blob.c
@@ -3517,7 +3517,7 @@ MagickExport MagickOffsetType SeekBlob(Image *image,
       break;
     case FileStream:
     {
-      if (fseek(image->blob->file,(long) offset,whence) < 0)
+      if (fseek(image->blob->file,offset,whence) < 0)
         return(-1);
       image->blob->offset=TellBlob(image);
       break;
